requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-heroku==0.3.1
django-rest-swagger==2.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
drf-yasg==1.12.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
oauthlib==2.1.0
openapi-codec==1.3.2
pbr==5.1.1
Pillow==5.4.1
pluggy==0.8.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
PyJWT==1.7.1
pylint==2.2.2
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.1.0
ruamel.yaml==0.15.86
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.0
